(PC-27674)[PRO] refactor: move header to page level
